app/admin/requirements.txt
azure-core==1.32.0
azure-identity==1.19.0
azure-monitor-opentelemetry==1.6.4
azure-storage-blob==12.23.1
streamlit==1.41.1
